+++ packages/fast-eslint/spec/fixtures/airbnb/bad.js
@@ -1,0 +1,12 @@
+var foo = 'bar';
+
+if(foo == bar) foo = '42'
+
+function myFunc() {
+  return 'doSmthg';
+}
+
+
+
+a
+myFunc();