+++ packages/autocomplete-paths/CHANGELOG.md
@@ -1,3 +1,22 @@
+# [2.17.0](https://github.com/atom-community/autocomplete-paths/compare/v2.16.0...v2.17.0) (2021-11-29)
+
+
+### Features
+
+* add setting to show image preview ([#188](https://github.com/atom-community/autocomplete-paths/issues/188)) ([a5546b9](https://github.com/atom-community/autocomplete-paths/commit/a5546b9d43a8adfc0d9f720b2f0f1c5ba1dc6782))
+
+# [2.16.0](https://github.com/atom-community/autocomplete-paths/compare/v2.15.2...v2.16.0) (2021-11-29)
+
+
+### Bug Fixes
+
+* update deps ([#279](https://github.com/atom-community/autocomplete-paths/issues/279)) ([5e49f36](https://github.com/atom-community/autocomplete-paths/commit/5e49f363fa0a2bbfad7446bd03c5618d7186f2e5))
+
+
+### Features
+
+* add setting to not follow symlinks ([#278](https://github.com/atom-community/autocomplete-paths/issues/278)) ([964faee](https://github.com/atom-community/autocomplete-paths/commit/964faee12d23380b40a739dd4c70805699e2e073))
+
 ## [2.15.2](https://github.com/atom-community/autocomplete-paths/compare/v2.15.1...v2.15.2) (2021-05-02)